+++ CHANGELOG.md
@@ -3,6 +3,10 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
+## [1.0.1] - 2020-06-06
+- Added first implementation of a kafka producer.
+- Added routes to publish message into kafka stream.
+
 ## [1.0.0] - 2020-05-25
 - Added file for use with docker-compose. Allows to run apache kafka in your local environment.